refactor: Refactor virtual environment managed in kraken-wrapper to make use of `uv pip freeze` instead of the homegrown `kraken query env` command, which is slower and more brittle.
